+++ src/main/windows/editor.js
@@ -1,7 +1,7 @@
 import path from 'path'
 import { BrowserWindow, dialog, ipcMain } from 'electron'
 import { enable as remoteEnable } from '@electron/remote/main'
-import log from 'electron-log'
+import log from 'electron-log/main'
 import windowStateKeeper from 'electron-window-state'
 import { isChildOfDirectory, isSamePathSync } from 'common/filesystem/paths'
 import BaseWindow, { WindowLifecycle, WindowType } from './base'